[KEYCLOAK-5798] Move keycloak-nodejs-auth-utils and keycloak-nodejs-connect repositories
* Merge of repositories to keep original authors
* Move child-repo/lib to middleware/auth-utils
* Removal of useless dependencies
* Inclusion/fix auth-utils integration and unit tests
* Changes at the keycloak server start up script to import a new realm
