--- child-repo/.gitignore
@@ -1,6 +1,0 @@
-.idea
-node_modules
-docs
-./keycloak*
-keycloak-*
-*.log
