--- child-repo/.jshintrc
@@ -1,5 +1,0 @@
-{
-  "esversion": 6,
-  "node": true,
-  "sub": true
-}
