--- child-repo/.npmignore
@@ -1,6 +1,0 @@
-node_modules
-*keycloak.json
-npm-debug.log
-.idea
-.grunt
-doc
