--- child-repo/.travis.yml
@@ -1,19 +1,0 @@
-sudo: required
-dist: trusty
-language: node_js
-node_js:
-  - "4"
-  - "6"
-  - "7"
-  - "8"
-before_script:
-   - ./scripts/start-server.sh
-script:
-   - make ci
-after_success:
-   - npm run coveralls
-after_script:
-   - ./scripts/stop-server.sh
-env:
-  global:
-    - USER: ci
