--- child-repo/scripts/stop-server.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-. scripts/version.sh
-
-${KEYCLOAK}/bin/jboss-cli.sh --connect command=:shutdown
