--- child-repo/test/integration/helper.js
@@ -1,12 +1,0 @@
-module.exports = {
-  dummyReply: {
-    'access_token': 'Dummy access token',
-    'expires_in': 2,
-    'refresh_expires_in': 1800,
-    'refresh_token': 'Dummy refresh token',
-    'token_type': 'bearer',
-    'id_token': 'Dummy id token',
-    'not-before-policy': 1462208947,
-    'session_state': '22e0b5bd-fb0f-4f99-93aa-a60c4b934c88'
-  }
-};
