+++ middleware/auth-utils/grant-manager.js
@@ -198,7 +198,7 @@ GrantManager.prototype.userInfo = function userInfo (token, callback) {
   options.headers = {
     'Authorization': 'Bearer ' + t,
     'Accept': 'application/json',
-    'X-Client': 'keycloak-nodejs-auth-utils'
+    'X-Client': 'keycloak-nodejs-connect'
   };
 
   const promise = new Promise((resolve, reject) => {
@@ -385,10 +385,10 @@ const postOptions = (manager, path) => {
   const opts = URL.parse(manager.realmUrl + realPath);
   opts.headers = {
     'Content-Type': 'application/x-www-form-urlencoded',
-    'X-Client': 'keycloak-nodejs-auth-utils'
+    'X-Client': 'keycloak-nodejs-connect'
   };
   if (!manager.public) {
-    opts.headers['Authorization'] = 'Basic ' + new Buffer(manager.clientId + ':' + manager.secret).toString('base64');
+    opts.headers.Authorization = 'Basic ' + new Buffer(manager.clientId + ':' + manager.secret).toString('base64');
   }
   opts.method = 'POST';
   return opts;
