+++ scripts/start-server.sh
@@ -35,6 +35,12 @@ fi
 
 # Start the server
 $KEYCLOAK/bin/add-user-keycloak.sh -u admin -p admin
-$KEYCLOAK/bin/standalone.sh -Djava.net.preferIPv4Stack=true > keycloak.log 2>&1 &
+
+# Start the server
+$KEYCLOAK/bin/standalone.sh -Djava.net.preferIPv4Stack=true \
+                            -Dkeycloak.migration.action=import \
+                            -Dkeycloak.migration.provider=singleFile \
+                            -Dkeycloak.migration.file=./test/fixtures/auth-utils/nodejs-test-realm.json \
+                            -Dkeycloak.migration.strategy=OVERWRITE_EXISTING > keycloak.log 2>&1 &
 
 waitForServer
