+++ test/fixtures/auth-utils/keycloak-confidential.json
