+++ test/fixtures/auth-utils/keycloak-with-public-key.json
